-- guillaumebogard in process

+++ src/main/scala/aa_cookbook/x002/Floats.scala
@@ -1,9 +1,9 @@
 package aa_cookbook.x002
 
 object Floats extends App {
-  val x1 = 0.1f
-  val x2 = 0.2f
-  val x3 = 0.3f
+  val x1 = 0.1
+  val x2 = 0.2
+  val x3 = 0.3
 
   if (x1 + x2 == x3) println("equals!")
     else println("no equals!")
@@ -11,14 +11,12 @@ object Floats extends App {
   def ~=(x: Double, y:Double, precision: Double=0.001) =
     if ((x-y).abs < precision) true else false
 
-  implicit class PreciseComparator(d: Float) {
-    def ~=(x: Float, precision: Float) = (d-x).abs < 0.01
+  implicit class PreciseComparator(d: Double) {
+    def ~=(x: Double, precision: Double) = (d-x).abs < 0.01
   }
 
   println(~=(x1+x2, x3, 0.001))
   println(~=(x1+x2, x3))
   ~=(x1+x2, x3)
   println(x1.~=(x2,0.01f))
-
-
 }
